refactor(ctterm): extract map helpers and denest key handling
Split map TUI helpers into focused modules and refactor diplomacy/development TUI key handling into smaller helpers to keep screens testable and maintainable.

Made-with: Cursor
